+++ rust/core/audio/interpreter/load.rs
@@ -1,0 +1,18 @@
+use crate::core::{
+    parser::statement::{ Statement, StatementKind },
+    shared::value::Value,
+    store::variable::VariableTable,
+};
+
+pub fn interprete_load_statement(
+    stmt: &Statement,
+    variable_table: &mut VariableTable
+) -> Option<VariableTable> {
+    if let StatementKind::Load { source, alias } = &stmt.kind {
+        variable_table.set(alias.to_string(), Value::String(source.clone()));
+
+        return Some(variable_table.clone());
+    }
+
+    None
+}